Move buddies in utils/
