+++ tests/integrations/crossed_integrations/test_kinesis.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 import json
 
-from tests.integrations.crossed_integrations.test_kafka import _python_buddy
+from utils.buddies import python_buddy
 from utils import interfaces, scenarios, weblog, missing_feature, features
 from utils.tools import logger
 
@@ -201,6 +201,6 @@ class _Test_Kinesis:
 @features.aws_kinesis_span_creationcontext_propagation_via_message_attributes_with_dd_trace
 class Test_Kinesis_PROPAGATION_VIA_MESSAGE_ATTRIBUTES(_Test_Kinesis):
     buddy_interface = interfaces.python_buddy
-    buddy = _python_buddy
+    buddy = python_buddy
     WEBLOG_TO_BUDDY_STREAM = "Test_Kinesis_propagation_via_message_attributes_weblog_to_buddy"
     BUDDY_TO_WEBLOG_STREAM = "Test_Kinesis_propagation_via_message_attributes_buddy_to_weblog"
