+++ tests/integrations/crossed_integrations/test_rabbitmq.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 import json
 
-from tests.integrations.crossed_integrations.test_kafka import _java_buddy
+from utils.buddies import java_buddy
 from utils import interfaces, scenarios, weblog, missing_feature, features
 from utils.tools import logger
 
@@ -234,7 +234,7 @@ class _Test_RabbitMQ:
 @features.rabbitmq_span_creationcontext_propagation_with_dd_trace
 class Test_RabbitMQ_Trace_Context_Propagation(_Test_RabbitMQ):
     buddy_interface = interfaces.java_buddy
-    buddy = _java_buddy
+    buddy = java_buddy
     WEBLOG_TO_BUDDY_QUEUE = "Test_RabbitMQ_Propagation_weblog_to_buddy"
     WEBLOG_TO_BUDDY_EXCHANGE = "Test_RabbitMQ_Propagation_weblog_to_buddy_exchange"
     WEBLOG_TO_BUDDY_ROUTING_KEY = "Test_RabbitMQ_Propagation_weblog_to_buddy_routing_key"
